requirements.txt
Flask==2.0.1
google-generativeai==0.3.2
PyPDF2==1.26.0
sumy==0.8.1
nltk==3.6.5
python-dotenv==0.19.1
Werkzeug==2.0.1
